Project cleanup and refactoring types to ModelDescriptor
