+++ Assets/Package/Editor/SketchfabBrowser.cs
@@ -8,6 +8,7 @@ using System.IO.Compression;
 using System.Linq;
 using System.Threading.Tasks;
 using Cysharp.Threading.Tasks;
+using Package.Runtime;
 using UnityEngine.Networking;
 using UnityEngine.Serialization;
 using UnityEngine.UI;
@@ -16,7 +17,7 @@ using Object = UnityEngine.Object;
 public class SketchfabBrowser : EditorWindow
 {
     private string modelId = "564e02a97528499388ca00d3c6bdb044";
-    private string apiToken; 
+    private string apiToken;
     private bool connected;
     private string accountName;
     public Model CurrentModel = new Model();
@@ -27,12 +28,12 @@ public class SketchfabBrowser : EditorWindow
     private bool moreInfo;
     private PageModels currentPageModels;
     private string searchKeyword = "milk";
-    
+
     public PageModels pageModels;
     private List<SearchThumb> searchThumbs = new();
     private bool isSearching;
     private const string SketchfabTokenKey = "SketchfabTokenKey";
-    
+
     private Vector2 scrollPosition;
     private GridPanel panelDrawer;
 
@@ -43,14 +44,16 @@ public class SketchfabBrowser : EditorWindow
     {
         SketchfabBrowser window = GetWindow<SketchfabBrowser>();
         window.titleContent = new GUIContent("Sketchfab Browser", window.windowIcon);
-        
+
     }
 
     private void OnEnable()
     {
-        windowIcon = (Texture2D)AssetDatabase.LoadAssetAtPath("Assets/Package/Editor/res/sketchfab-logo.png", typeof(Texture2D));
+        windowIcon =
+            (Texture2D)AssetDatabase.LoadAssetAtPath("Assets/Package/Editor/res/sketchfab-logo.png", typeof(Texture2D));
         titleContent.image = windowIcon;
-        apiToken = PlayerPrefs.GetString(SketchfabTokenKey, "your-sketchfab-api-token"); //https://sketchfab.com/settings/password
+        apiToken = PlayerPrefs.GetString(SketchfabTokenKey,
+            "your-sketchfab-api-token"); //https://sketchfab.com/settings/password
         panelDrawer = new GridPanel();
         Instance = this;
     }
@@ -65,7 +68,7 @@ public class SketchfabBrowser : EditorWindow
             GUILayout.Label($"Unavailable in play mode.", EditorStyles.boldLabel);
             return;
         }
-        
+
         if (!connected)
         {
             DrawConnectionUI();
@@ -73,8 +76,9 @@ public class SketchfabBrowser : EditorWindow
         else
         {
             GUILayout.Label($"Connected as {accountName}", EditorStyles.boldLabel);
-            
+
             #region ٍSearch Button
+
             GUILayout.Space(20);
 
             string searchKewordInput = "SearchKeywordInput";
@@ -85,20 +89,22 @@ public class SketchfabBrowser : EditorWindow
             {
                 Search24(searchKeyword).Forget();
             }
+
             GUI.enabled = true;
 
             if (GUI.GetNameOfFocusedControl() == searchKewordInput && Event.current.keyCode == KeyCode.Return)
             {
                 Search24(searchKeyword).Forget();
             }
+
             #endregion
-            
+
             if (pageModels.results.Length > 0)
             {
                 DisplayResults();
                 return;
             }
-            
+
             GUILayout.Label("No models available.");
         }
     }
@@ -110,78 +116,7 @@ public class SketchfabBrowser : EditorWindow
         EditorGUILayout.EndScrollView();
     }
 
-    private void DrawModelDetails()
-    {
-        { 
-            GUILayout.Space(20);
-            modelId = EditorGUILayout.TextField("Model ID", modelId);
 
-            if (string.IsNullOrEmpty(modelId) || modelId.Length < 13)
-            {
-                GUI.enabled = false;
-            }
-
-            if (GUILayout.Button("Fetch Model Details"))
-            {
-                FetchInfo().Forget();
-                Repaint();
-            }
-
-            GUI.enabled = true;
-            
-            if (CurrentModel != null)
-            {
-
-                var dd = DateTime.ParseExact(CurrentModel.updatedAt, "yyyy-MM-dd'T'HH:mm:ss", CultureInfo.InvariantCulture);
-                string updateShortDate = dd.ToString("dd-MMMM-yyyy");
-                
-                string desc = CurrentModel.description.Length > 20 ? CurrentModel.description.Substring(0, 20): CurrentModel.description;
-                EditorGUILayout.HelpBox($"Model Name: {CurrentModel.name}\nDescription: {desc}\n"
-                                        + $"Updated at: {updateShortDate}"
-                                        , MessageType.Info);
-                //Load thumbnail only once
-                if (thumb == null) GetThumbnail().Forget();
-                else GUILayout.Label(thumb, GUILayout.Width(256));
-
-                //Download model
-                if (CurrentModel.isDownloadable)
-                {
-                    GUILayout.Label($"License: {CurrentModel.license.label}", EditorStyles.boldLabel);
-                    if (GUILayout.Button("Download Model")) DownloadModel(modelId, CurrentModel.name).Forget();
-                }
-                else
-                {
-                    var price = (CurrentModel.price/100).ToString("N2");
-                    GUILayout.Label($"Price: {price}", EditorStyles.boldLabel);
-                    
-                }
-                
-                GUILayout.Space(10);
-            }
-            
-            moreInfo = EditorGUILayout.BeginFoldoutHeaderGroup(moreInfo, "More info", EditorStyles.foldoutHeader);
-
-            if (moreInfo && CurrentModel != null)
-            {
-                GUILayout.Label($"Mesh Data", EditorStyles.boldLabel);
-                EditorGUI.indentLevel += 50;
-                
-                GUILayout.Label($"Vertex Count: {CurrentModel.vertexCount.ToString("N0")}");
-                GUILayout.Label($"Face Count: {CurrentModel.faceCount.ToString("N0")}");
-                EditorGUI.indentLevel -= 50;
-                GUILayout.Label($"Materials: {CurrentModel.materialCount}");
-                GUILayout.Label($"Textures: {CurrentModel.textureCount}");
-                
-                if(CurrentModel.animationCount > 0)
-                {
-                    GUILayout.Label($"Animations: {CurrentModel.animationCount.ToString("N0")}");
-                }
-            }
-            
-            GUILayout.Space(10);
-            if (GUILayout.Button("Goto Model", hyperlinkStyle)) Application.OpenURL(CurrentModel.viewerUrl);
-        }
-    }
 
     private void DrawConnectionUI()
     {
@@ -208,7 +143,7 @@ public class SketchfabBrowser : EditorWindow
         }
 
         GUI.enabled = true;
-        
+
         if (GUI.GetNameOfFocusedControl() == tokenFieldName && Event.current.keyCode == KeyCode.Return)
         {
             GUI.FocusControl(null);
@@ -219,7 +154,7 @@ public class SketchfabBrowser : EditorWindow
     private async UniTaskVoid GetThumbnail()
     {
         thumb = await DownloadImage(CurrentModel.thumbnails.images[0].url);
-         Repaint();
+        Repaint();
     }
 
     async UniTaskVoid ConnectToSketchfab()
@@ -242,7 +177,7 @@ public class SketchfabBrowser : EditorWindow
                 AccountInfo accountInfo = JsonUtility.FromJson<AccountInfo>(request.downloadHandler.text);
                 accountName = accountInfo.username;
                 Repaint();
-                
+
                 PlayerPrefs.SetString(SketchfabTokenKey, token);
                 PlayerPrefs.Save();
             }
@@ -257,12 +192,13 @@ public class SketchfabBrowser : EditorWindow
     async UniTaskVoid FetchInfo() => await FetchModelInfo(modelId);
     async UniTaskVoid Download() => await DownloadModel(modelId, CurrentModel.name);
 
-    async UniTask Search24(string keyword, string after= null)
+    async UniTask Search24(string keyword, string after = null)
     {
         isSearching = true;
         GUI.FocusControl(null);
 
-        string searchRequest = $"https://api.sketchfab.com/v3/search?type=models&downloadable=true&purchasable=true&tags={keyword}&name={keyword}&description={keyword}&sort_by=-likeCount&per_page=24&after={after}";
+        string searchRequest =
+            $"https://api.sketchfab.com/v3/search?type=models&downloadable=true&purchasable=true&tags={keyword}&name={keyword}&description={keyword}&sort_by=-likeCount&per_page=24&after={after}";
         using (var request = UnityWebRequest.Get(searchRequest))
         {
             request.SetRequestHeader("Authorization", $"Token {apiToken}");
@@ -306,6 +242,7 @@ public class SketchfabBrowser : EditorWindow
                 }
             }
         }
+
         isSearching = false;
     }
 
@@ -366,26 +303,29 @@ public class SketchfabBrowser : EditorWindow
             request.SetRequestHeader("Authorization", $"Token {apiToken}");
             await request.SendWebRequest();
 
-            if (request.result == UnityWebRequest.Result.ConnectionError || request.result == UnityWebRequest.Result.ProtocolError)
+            if (request.result == UnityWebRequest.Result.ConnectionError ||
+                request.result == UnityWebRequest.Result.ProtocolError)
             {
                 Debug.LogError("Error: " + request.error);
             }
             else if (request.result == UnityWebRequest.Result.Success)
             {
-                if(request.responseCode == 401)
+                if (request.responseCode == 401)
                 {
                     Debug.LogError("Unauthorized: Invalid API token.");
                 }
                 else
                 {
-                    ModelDownloadInfo downloadInfo = JsonUtility.FromJson<ModelDownloadInfo>(request.downloadHandler.text);
-                    
-                    
+                    ModelDownloadInfo downloadInfo =
+                        JsonUtility.FromJson<ModelDownloadInfo>(request.downloadHandler.text);
+
+
                     await DownloadModelFromUrl(downloadInfo.gltf.url, ModelFormatExtension.gltf);
                     //await DownloadModelFromUrl(downloadInfo.source.url, ModelFormatExtension.fbx);
                 }
             }
         }
+
         IsDownloading = false;
         CurrentModel.IsDownloading = false;
         Repaint();
@@ -402,7 +342,7 @@ public class SketchfabBrowser : EditorWindow
             {
                 byte[] modelBytes = request.downloadHandler.data;
 
-                string fileName = CurrentModel.name + $".zip"; 
+                string fileName = CurrentModel.name + $".zip";
                 //string savePath = EditorUtility.SaveFilePanel("Save Model", "", modelId, "");
                 Directory.CreateDirectory(defaultSavePath);
                 string savePath = Path.Combine(defaultSavePath, fileName);
@@ -430,10 +370,7 @@ public class SketchfabBrowser : EditorWindow
 
         try
         {
-            await UniTask.Run(() =>
-            {
-                ZipFile.ExtractToDirectory(savePath, unpackPath);
-            });
+            await UniTask.Run(() => { ZipFile.ExtractToDirectory(savePath, unpackPath); });
             AssetDatabase.Refresh(); //to avoid warnings Import Error Code(4)
         }
         catch (IOException e)
@@ -443,7 +380,7 @@ public class SketchfabBrowser : EditorWindow
 
         File.Delete(savePath);
         Debug.Log($"Model is downloaded to: " + defaultSavePath);
-        
+
         var targetFilePath = Directory.EnumerateFiles(unpackPath).FirstOrDefault();
         if (targetFilePath != null)
         {
@@ -455,97 +392,6 @@ public class SketchfabBrowser : EditorWindow
             }
         }
     }
-
-    [System.Serializable]
-    public class PageModels
-    {
-        public Model[] results;
-        public Pagination pagination;
-    }
-    
-    [System.Serializable]
-    public class Pagination
-    {
-        public string next;
-    }
-
-    [System.Serializable]
-    public class Model
-    {
-        public string name;
-        public string description;
-        public string uri; // URL of the model's page
-        public Thumbnails thumbnails; // URLs and images of the model's thumbnails
-        public bool isDownloadable;
-        public string viewerUrl;
-        public int vertexCount;
-        public int textureCount;
-        public int materialCount;
-        public int faceCount;
-        public int animationCount;
-        
-        public float price;
-        public string updatedAt;
-        public License license;
-        public string uid;
-        public bool IsDownloading;
-    }
-
-    [System.Serializable]
-    public class Thumbnails
-    {
-        public Thumbnail[] images;
-
-        [System.Serializable]
-        public class Thumbnail
-        {
-            public string url;
-        }
-    }
-
-    [System.Serializable]
-    public class License
-    {
-        public string label;
-    }
-
-    [System.Serializable]
-    private class AccountInfo
-    {
-        public string username;
-    }
-    
-    [System.Serializable]
-    public class ModelDownloadInfo
-    {
-        public ModelFormat gltf;
-        [FormerlySerializedAs("fbx")] public ModelFormat source;
-
-        [System.Serializable]
-        public class ModelFormat
-        {
-            public string url;
-            public int size;
-            public int expires;
-        }
-    }
-
-    public enum ModelFormatExtension
-    {
-        gltf,
-        fbx
-    }
 }
 
-public class SearchThumb
-{
-    public readonly string uid;
-    public readonly Texture2D thumb;
-
-    public SearchThumb(string modelUid, Texture2D texture2D)
-    {
-        uid = modelUid;
-        thumb = texture2D;
-    }
-}
-
+
