+++ Assets/Package/Runtime/ModelDescriptor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46b1777801804ab4e998473fe75e2e04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
